split into blocks
